requirements.txt
python-telegram-bot
typing
deep_translator
torch
langchain
chromadb